Chore: Reintroduce 2FA functionality with event handlers for enabling and disabling 2FA
